+++ pcbnew/widgets/search_handlers.h
@@ -39,7 +39,7 @@ public:
 
     wxString GetResultCell( int aRow, int aCol ) override
     {
-        if( aRow >= m_hitlist.size() )
+        if( aRow >= static_cast<int>(m_hitlist.size() ) )
             return wxEmptyString;
 
         BOARD_ITEM* item = m_hitlist[aRow];
